Feat: implement albumart command, make covercache optional #145

+++ src/mpd_client/mpd_client_cover.c
@@ -23,45 +23,76 @@
 #include "mpd_client_utility.h"
 #include "mpd_client_cover.h"
 
-//private definitions
-static int mpd_client_albumart(t_mpd_state *mpd_state, const char *uri, int fd);
-
-//public functions
 sds mpd_client_getcover(t_config *config, t_mpd_state *mpd_state, sds buffer, sds method, int request_id,
-                     const char *uri) 
+                        const char *uri, sds *binary)
 {
-    sds tmp_file = sdscatfmt(sdsempty(), "%s/covercache/%s.XXXXXX", config->varlibdir, uri);
-    int fd = mkstemp(tmp_file);
-    if (fd < 0 ) {
-        LOG_ERROR("Can't open %s for write", tmp_file);
-        sdsfree(tmp_file);
-        buffer = jsonrpc_respond_message(buffer, method, request_id, "Can't write covercache file", true);
-        return buffer;
-    }
+    unsigned offset = 0;
     unsigned size = 0;
-    if (mpd_state->feat_mpd_albumart == true) {
-        size = mpd_client_albumart(mpd_state, uri, fd);
-    }
-    if (size == 0 && mpd_state->feat_mpd_readpicture == true) {
-      //todo: implement readpicture command
-    }
-
-    if (size > 0) {
-        sds mime_type = get_mime_type_by_magic(tmp_file);
-        sds ext = get_ext_by_mime_type(mime_type);        
-        sds cover_file = sdscatfmt(sdsempty(), "%s/covercache/%s.%s", config->varlibdir, uri, ext);
-        if (rename(tmp_file, cover_file) == -1) {
-            LOG_ERROR("Rename file from %s to %s failed", tmp_file, cover_file);
-            buffer = jsonrpc_respond_message(buffer, method, request_id, "Can't write covercache file", true);
+    FILE *fp = NULL;
+    sds tmp_file = sdscatfmt(sdsempty(), "%s/covercache/%s.XXXXXX", config->varlibdir, uri);
+    if (config->readonly == false) {
+        int fd = mkstemp(tmp_file);
+        if (fd < 0 ) {
+            LOG_ERROR("Can't open %s for write", tmp_file);
             sdsfree(tmp_file);
-            sdsfree(cover_file);
-            sdsfree(mime_type);
-            sdsfree(ext);
+            buffer = jsonrpc_respond_message(buffer, method, request_id, "Can't write covercache file", true);
             return buffer;
         }
+        fp = fdopen(fd, "w");
+    }
+#ifdef EMBEDDED_LIBMPDCLIENT
+    if (mpd_state->feat_mpd_albumart == true) {
+        struct mpd_albumart albumart_buffer;
+        struct mpd_albumart *albumart;
+        while ((albumart = mpd_run_albumart(mpd_state->conn, uri, offset, &albumart_buffer)) != NULL) {
+            if (config->readonly == false) {
+                fwrite(albumart->data, 1, albumart->data_length, fp);
+            }
+            if (albumart->size < 10485760 || config->readonly == true) {
+                // smaller than 10 MB, or readonly mode
+                *binary = sdscatlen(*binary, albumart->data, albumart->data_length);
+            }
+            offset += albumart->data_length;
+            size = albumart->size;
+        }
+    }
+    if (size == 0 && mpd_state->feat_mpd_readpicture == true) {
+        //todo: implement readpicture command
+    }
+#endif
+    if (config->readonly == false) {
+        fclose(fp);
+    }
+    if (size > 0) {
+        sds mime_type;
+        if (sdslen(*binary) > 0) {
+            mime_type = get_mime_type_by_magic_stream(*binary);
+        }
+        else {
+            mime_type = get_mime_type_by_magic(tmp_file);
+        }
+        sds ext = get_ext_by_mime_type(mime_type);        
+        sds cover_file = sdscatfmt(sdsempty(), "%s/covercache/%s.%s", config->varlibdir, uri, ext);
+        if (config->readonly == false) {
+            if (rename(tmp_file, cover_file) == -1) {
+                LOG_ERROR("Rename file from %s to %s failed", tmp_file, cover_file);
+                sdsfree(mime_type);
+                sdsfree(ext);
+                sdsfree(cover_file);
+                buffer = jsonrpc_respond_message(buffer, method, request_id, "Can't write covercache file", true);
+                return buffer;
+            }
+        }
+
         buffer = jsonrpc_start_result(buffer, method, request_id);
         buffer = sdscat(buffer, ",");
-        buffer = tojson_char(buffer, "coverfile", cover_file, true);
+        if (sdslen(*binary) > 0) {
+            buffer = tojson_char(buffer, "coverfile", "binary", true);
+        }
+        else {
+            buffer = tojson_char(buffer, "coverfile", "file", true);
+        }
+        buffer = tojson_char(buffer, "coverfile_name", cover_file, true);
         buffer = tojson_char(buffer, "mime_type", mime_type, false);
         buffer = jsonrpc_end_result(buffer);
         sdsfree(mime_type);
@@ -69,30 +100,11 @@ sds mpd_client_getcover(t_config *config, t_mpd_state *mpd_state, sds buffer, sd
         sdsfree(cover_file);
     }
     else {
-        unlink(tmp_file);
-        buffer = jsonrpc_respond_message(buffer, method, request_id, "No albumart found by mpd", true);
+        buffer = jsonrpc_respond_message(buffer, method, request_id, "No albumart found by mpd", true);    
+        if (config->readonly == false) {
+            unlink(tmp_file);
+        }
     }
     sdsfree(tmp_file);
     return buffer;
-}
-
-//private functions
-static int mpd_client_albumart(t_mpd_state *mpd_state, const char *uri, int fd) {
-    unsigned size = 0;
-#ifdef EMBEDDED_LIBMPDCLIENT
-    struct mpd_albumart albumart_buffer;
-    struct mpd_albumart *albumart;
-    unsigned offset = 0;
-    FILE *fp = fdopen(fd, "w");
-    while ((albumart = mpd_run_albumart(mpd_state->conn, uri, offset, &albumart_buffer)) != NULL) {
-        fwrite(albumart->data, 1, albumart->data_length, fp);
-	offset += albumart->data_length;
-	size = albumart->size;
-    }
-    fclose(fp);
-#else
-    (void) uri;
-    (void) fd;
-#endif
-    return size;
 }
